requirements.txt
streamlit==0.81.0
Pillow==8.0.1
requests==2.25.0
requests-oauthlib==1.3.0